+++ src/SQLBaseModifier.php
@@ -1,6 +1,18 @@
 <?php
 class SQLBaseModifier
 {
+  public static function default_handler($modifier, &$queryObject, $field) {
+    $fieldName = $queryObject['select'][$field]['field'];
+    $aliasName = $queryObject['select'][$field]['alias'];
+
+    $queryObject['select']["$modifier($fieldName)" . ($aliasName ? "@$aliasName" : '')] = [
+      'field' => $fieldName,
+      'alias' => $aliasName,
+    ];
+
+    unset($queryObject['select'][$field]);
+  }
+
   public static function mod_asc(&$queryObject, $field) {
     $queryObject['order'][] = [
       'field' => $queryObject['select'][$field]['field'],
@@ -23,17 +35,5 @@ class SQLBaseModifier
       $name = $queryObject['select'][$field]['modifier']['group'][0];
       $queryObject['having'][] = "$group = $name";
     }
-  }
-
-  public static function mod_count(&$queryObject, $field) {
-    $fieldName = $queryObject['select'][$field]['field'];
-    $aliasName = $queryObject['select'][$field]['alias'];
-
-    $queryObject['select']["count($fieldName)" . ($aliasName ? "@$aliasName" : '')] = [
-			'field' => "$fieldName",
-			'alias' => $aliasName,
-		];
-
-    unset($queryObject['select'][$field]);
   }
 }